feat: 🎸 upgrade spacing component
