+++ packages/ui/src/components/Banner/BannerStyles.ts
@@ -3,7 +3,7 @@ import {createStyles, makeStyles, Theme} from "@material-ui/core/styles";
 const BannerStyles = makeStyles((theme: Theme) =>
   createStyles({
     banner: {
-      padding: theme.spacing(12, 0, 6),
+      padding: theme.spacing(12, 0, 0),
     },
     item: {
       display: "flex",
