+++ packages/ui/src/components/Product/ProductStyles.ts
@@ -3,11 +3,10 @@ import {createStyles, makeStyles, Theme} from "@material-ui/core/styles";
 const ProductStyles = makeStyles((theme: Theme) =>
   createStyles({
     product: {
-      padding: theme.spacing(3, 3, 3),
+      padding: theme.spacing(0, 3, 0),
     },
     root: {
       flexGrow: 1,
-      padding: theme.spacing(8, 0, 6),
     },
   }),
 );
