+++ packages/ui/src/components/Section/SectionStyles.tsx
@@ -3,7 +3,7 @@ import {createStyles, makeStyles, Theme} from "@material-ui/core/styles";
 const SectionStyles = makeStyles((theme: Theme) =>
   createStyles({
     section: {
-      padding: theme.spacing(9, 3, 3),
+      padding: theme.spacing(0, 3, 0),
     },
     title: {
       // backgroundColor: theme.palette.primary.main,
