+++ packages/ui/src/components/Spacing/Spacing.tsx
@@ -1,0 +1,6 @@
+import Box from "@material-ui/core/Box";
+import React from "react";
+
+export default function Spacing(): JSX.Element {
+  return <Box py={5} />;
+}
