+++ packages/ui/src/components/Spacing/index.ts
@@ -1,0 +1,1 @@
+export {default} from "./Spacing";
