+++ packages/ui/src/index.ts
@@ -8,6 +8,7 @@ export {default as Pricing} from "./components/Pricing";
 export {default as Product} from "./components/Product";
 export {default as ProTip} from "./components/ProTip";
 export {default as Section} from "./components/Section";
+export {default as Spacing} from "./components/Spacing";
 export {default as Testimonial} from "./components/Testimonial";
 export {default as TestimonialCard} from "./components/TestimonialCard";
